requirements.txt
Django==4.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-guardian==2.4.0
django-userforeignkey==0.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-yasg==1.21.7
google-auth==2.34.0
mysqlclient==2.2.4
pillow==10.4.0
pyjwt==2.9.0
pytz==2024.1
requests==2.32.3
rest-framework-generic-relations==2.1.0
sorl-thumbnail==12.10.0
